refactor verify folder
