+++ lmfdb/verify/mf/mf_gamma1_portraits.py
