+++ lmfdb/verify/mf/mf_hecke_cc.py
@@ -6,7 +6,7 @@ from lmfdb.lmfdb_database import db
 from lmfdb.backend.utils import IdentifierWrapper as Identifier
 from psycopg2.sql import SQL, Literal
 from .mf import MfChecker
-from .verification import overall, overall_long, slow
+from ..verification import overall, overall_long, slow
 
 class mf_hecke_cc(MfChecker):
     table = db.mf_hecke_cc
