+++ lmfdb/verify/mf/mf_hecke_newspace_traces.py
