+++ lmfdb/verify/mf/mf_hecke_traces.py
