+++ lmfdb/verify/mf/mf_newform_portraits.py
