+++ lmfdb/verify/mf/mf_subspaces.py
@@ -1,7 +1,7 @@
 
 from lmfdb import db
 from .mf import SubspacesChecker
-from .verification import overall
+from ..verification import overall
 
 class mf_subspaces(SubspacesChecker):
     table = db.mf_subspaces
